update to parse built in fields
We recently decided to get rid of a very old parse-server local fork in favor of the latest community release. We ran into an issue where one of the collections had a predefined lastUsed field configured as a string so we can store unix timestamps. That didn't play well with the latest release because lastUsed is considered a "reserved word". This caused many errors because the codebase is trying to treat it as a date object. We were able to fix this issue by deleting the field. But I think that parse reserved words need to be more specific for backward compatibility. 

* note there are probably other files that touch these fields.